# backend/docs/openapi-pretty.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/users/{userId}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "operationId" : "getUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "user-controller" ],
        "operationId" : "updateUserProfile",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserProfileUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/reactivate" : {
      "put" : {
        "tags" : [ "user-controller" ],
        "operationId" : "reactivateUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups/{groupId}" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "group-controller" ],
        "operationId" : "updateGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GroupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "group-controller" ],
        "operationId" : "deleteGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "deletedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/groups/{groupId}/archive" : {
      "put" : {
        "tags" : [ "group-controller" ],
        "operationId" : "archiveGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "archivedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{userId}/suspend" : {
      "put" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "suspendUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "reason",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}/reset-password" : {
      "put" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "resetUserPassword",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}/reactivate" : {
      "put" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "reactivateUser_1",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}/gdpr-delete" : {
      "put" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "deleteUserDataForCompliance",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "legalBasis",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/groups/{groupId}/moderate" : {
      "put" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "moderateGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "reason",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups" : {
      "post" : {
        "tags" : [ "group-controller" ],
        "operationId" : "createGroup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GroupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}/stats" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "operationId" : "getUserStats",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/search" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "operationId" : "searchUsers",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups/{groupId}/stats" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getGroupStats",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupStats"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups/user/{userId}" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getUserGroups",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/GroupResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users" : {
      "get" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "getAllUsers",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}" : {
      "get" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "getUserById",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}/groups" : {
      "get" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "getUserGroupsForSupport",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/GroupResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}/export" : {
      "get" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "exportUserData",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/groups" : {
      "get" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "getAllGroups",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageGroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/groups/{groupId}" : {
      "get" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "getGroupById",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UserProfileUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "phoneNumber" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          },
          "profilePictureUrl" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          }
        }
      },
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "profilePictureUrl" : {
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastLoginAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "GroupResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "createdBy" : {
            "$ref" : "#/components/schemas/UserResponse"
          },
          "defaultCurrency" : {
            "type" : "string"
          },
          "privacyLevel" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "PRIVATE", "SECRET" ]
          },
          "invitationPolicy" : {
            "type" : "string",
            "enum" : [ "CREATOR_ONLY", "ADMIN_ONLY", "ALL_MEMBERS" ]
          },
          "allowExternalPayments" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "ARCHIVED", "DELETED" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "members" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            },
            "uniqueItems" : true
          },
          "admins" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            },
            "uniqueItems" : true
          },
          "groupImageUrl" : {
            "type" : "string"
          }
        }
      },
...
      "GroupRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "description" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "createdByUserId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "updatedByUserId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "privacyLevel" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "PRIVATE", "SECRET" ]
          },
          "defaultCurrency" : {
            "type" : "string",
            "maxLength" : 3,
            "minLength" : 3
          }
        },
        "required" : [ "createdByUserId", "defaultCurrency", "name" ]
      },
...
      "SortObject" : {
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
      "GroupStats" : {
        "type" : "object",
        "properties" : {
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "memberCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "adminCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalExpenses" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalAmount" : {
            "type" : "number"
          },
          "settledAmount" : {
            "type" : "number"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "unsettledAmount" : {
            "type" : "number"
          },
          "settlementPercentage" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
...
      "PageUserResponse" : {
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}